separate out fetch into its own hook
